fix(json-schema): emit (float) cast instead of deprecated (double)
DenormalizerGenerator builds the int-to-float cast as Expr\Cast\Double
without a kind, which php-parser prints as (double). PHP 8.5 deprecates
the (double) and (real) cast syntax, so every generated client with
float properties emits deprecation warnings at runtime.

Set kind to KIND_FLOAT so the generated code reads (float). Regenerate
the expected fixtures and apply the same substitution to the package's
own committed normalizers, which were generated by this same code.

+++ src/Component/OpenApi2/Tests/fixtures/issue-770/expected/Normalizer/AprulesGpsCoordinatesNormalizer.php
@@ -38,13 +38,13 @@ class AprulesGpsCoordinatesNormalizer implements DenormalizerInterface, Normaliz
             return new Reference($data['$recursiveRef'], $context['document-origin']);
         }
         if (\array_key_exists('latitude', $data) && \is_int($data['latitude'])) {
-            $data['latitude'] = (double) $data['latitude'];
+            $data['latitude'] = (float) $data['latitude'];
         }
         if (\array_key_exists('longitude', $data) && \is_int($data['longitude'])) {
-            $data['longitude'] = (double) $data['longitude'];
+            $data['longitude'] = (float) $data['longitude'];
         }
         if (\array_key_exists('distance', $data) && \is_int($data['distance'])) {
-            $data['distance'] = (double) $data['distance'];
+            $data['distance'] = (float) $data['distance'];
         }
         if (\array_key_exists('latitude', $data)) {
             $object->setLatitude($data['latitude']);
